% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atomRecognition.R
\name{atomicRecognition}
\alias{atomicRecognition}
\alias{symb.Atoms}
\title{Atomic Recognition}
\usage{
\method{symb}{Atoms}(x, na.as.dummy = FALSE)
}
\arguments{
\item{x}{an object of class \sQuote{Atoms}}

\item{na.as.dyummy}{a logical value. Whether to consider NA values as dummy
atoms or not.}
}
\value{
a character vector. Atomic symbols.
}
\description{
Determine atomic symbols from an \sQuote{Atoms} object.
}
\details{
This function apply the \code{symb} method to the \code{x$atmtype} component 
of an \sQuote{Atoms} object.\cr If \code{x$atmtype} is \code{NULL}, 
\code{x$atmname} is used instead.\cr If \code{x$atmtype} contains empty 
character strings, they are replaced by their associated \code{x$atmname} 
values.
}
\keyword{manip}

