% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atoms.R
\name{Atoms}
\alias{Atoms}
\title{Create an object of class \sQuote{Atoms}}
\usage{
Atoms(topo.file, trj.files, selection = -1L, first = 1L, last = -1L,
  stride = 1L, topo.fmt = NULL, trj.fmt = NULL)
}
\arguments{
\item{topo.file}{a character vector. Topology file name. See details.}

\item{trj.files}{a character vector. Trajectory file name(s). See details.}

\item{selection}{an integer vector. Indexes of atoms to be read.}

\item{first}{an integer. First frame to be read.}

\item{last}{an integer. last frame to be read.}

\item{stride}{an integer. Step used for reading the trajectory frames.}

\item{topo.fmt}{a character string. Format used for reading topology file. 
See details}

\item{trj.fmt}{a character string. Format used for reading trajectory file. 
See details}
}
\description{
This function read topology and trajectory files to build an object of class 
\sQuote{Atoms}.
}
\details{
To build an object of class \sQuote{Atoms} a topology and trajectory file are
required. A batch of files can also also be provided for the trajectory. For 
the sake of simplicity, when \code{topo.file}/\code{trj.files} is missing, 
the function attempts to read the topology/trajectory from 
\code{trj.files}/\code{topo.file}.

By default, topology and trajectory file formats are guessed from file
extensions. Otherwise, they can be specified with the \code{topo.fmt} and 
\code{trj.fmt} arguments. Currently, the following file formats can be read: 
XYZ, PDB (ENT), RES, DCD, NC (NETCDF, NCDF) and LAMMPS.

The reading can be restricted to specific atoms and frames by using the 
\code{selection}, \code{first}, \code{last} and \code{stride} arguments.
}

