% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\docType{methods}
\name{cell}
\alias{cell}
\alias{cell,Atoms-method}
\alias{cell<-}
\alias{cell<-,Atoms-method}
\title{Lattice Parameters/Vectors}
\usage{
cell(x, ...)

cell(x, ...) <- value

\S4method{cell}{Atoms}(x, multi = FALSE, ...)

\S4method{cell}{Atoms}(x, multi = FALSE, ...) <- value
}
\arguments{
\item{x}{an R object containing lattice parameters/vectors to 
extracted/replaced.}

\item{value}{a 3x3 numeric matrix containing by column the coordinates of the
lattice vectors (when \code{multi == FALSE}) or a data.frame containing by
column the lattice parameters (when \code{multi == TRUE}) to be replaced.
In this case, columns must be named \code{a}, \code{b}, \code{c},
\code{alpha}, \code{beta} or \code{gamma}.}

\item{multi}{a logical value. Whether to extract/replace the lattice 
parameters/vectors of all the frames of the trajectory.}
}
\value{
When \code{multi = FALSE}, \code{cell} returns a 3x3 numeric matrix 
  containing by column the coordinates of the lattice vectors of the current 
  frame. Otherwise, a data.frame containing the lattice parameters \code{a}, 
  \code{b}, \code{c}, \code{alpha}, \code{beta} and \code{gamma} of each 
  frame of the trajectory is returned.
  
  \code{cell<-} returns an object of the same class as \code{x} with updated 
  lattice parameters/vectors.
}
\description{
Extract or replace lattice parameters/vectors.
}
\details{
When \code{multi == FALSE}, \code{cell} extracts the lattice vectors of the 
current frame. Otherwise, the lattice parameters of all the frames are 
extracted.

When \code{multi == FALSE}, \code{cell<-} replace the lattice vectors of the 
current frame. Otherwise, the lattice parameters of all the frames are 
replaced.
}
\note{
Lengths are in Angstrom and angles in degree.
}
\seealso{
\code{\link{current}} \code{\link{cell2mat}}
}

