% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellProperties.R
\docType{methods}
\name{cellProperties}
\alias{cellDensity,Atoms-method}
\alias{cellProperties}
\alias{cellVolume,Atoms-method}
\title{Volume and Density of a Simulation Cell}
\usage{
cellVolume(x, ...)

\S4method{cellVolume}{Atoms}(x, multi = FALSE, ...)

cellDensity(x, ...)

\S4method{cellDensity}{Atoms}(x, mass = x$mass, multi = FALSE, ...)
}
\arguments{
\item{x}{an R object}

\item{multi}{a logical value. Whether to calculate the volume/density of the 
simulation cell for all the frames of the trajectory.}

\item{mass}{a numeric vector. atomic masses. If \code{x$mass} is \code{NULL},
function \code{mass} is called to guess masses from atomic symbols.}
}
\value{
\code{cellVolume} returns a numeric vector containing the volume of
  the simulation cell.
  
  \code{cellDensity} returns a numeric vector containing the density of the
  simulation cell.
}
\description{
Calculate the volume or density of a simulation cell.
}
\note{
The volume is in Angstrom^3 cube and the density in g/cm^2.
}

