% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current.R
\name{current}
\alias{current}
\title{Current Frame}
\usage{
current(x)
}
\arguments{
\item{x}{an R object.}

\item{value}{an integer. Index of the current frame.}
}
\value{
\code{current} returns an integer.
  
  \code{current<-} returns an object of the same class as \code{x} with the
  current frame index updated.
}
\description{
Get the index of the current frame.
}

