% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractional.R
\name{fractional}
\alias{fractional}
\alias{fractional.Atoms}
\title{Fraction Coordinates}
\usage{
fractional(x, ...)

\method{fractional}{Atoms}(x, frame = current(x), ...)
}
\arguments{
\item{x}{an R object.}

\item{frame}{an integer. The frame for which to extract the fractional
coordinates.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A data.frame containing by columns the xyz fractional coordinates.
}
\description{
Extract fractional coordinates of a given frame of a trajectory.
}

