% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/freeVolume.R
\name{freeVolume.Atoms}
\alias{freeVolume}
\alias{freeVolume.Atoms}
\alias{freeVolume2}
\alias{freeVolume2.Atoms}
\title{freeVolume}
\usage{
\method{freeVolume}{Atoms}(x, radius = NULL)

\method{freeVolume2}{Atoms}(x, radius = NULL)

freeVolume(x, ...)

freeVolume2(x, ...)
}
\description{
freeVolume
}

