% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/guessTopology.R
\name{guessAngles.Atoms}
\alias{guessAngles}
\alias{guessAngles.Atoms}
\alias{guessBonds}
\alias{guessBonds.Atoms}
\alias{guessDihedrals}
\alias{guessDihedrals.Atoms}
\alias{guessImpropers}
\alias{guessImpropers.Atoms}
\alias{guessTopology}
\alias{guessTopology.Atoms}
\title{Guess Topology}
\usage{
\method{guessAngles}{Atoms}(x)

\method{guessBonds}{Atoms}(x, radius = NULL, safety = 1.2, usePBC = TRUE)

\method{guessDihedrals}{Atoms}(x)

\method{guessImpropers}{Atoms}(x)

guessBonds(x, ...)

guessAngles(x)

guessDihedrals(x)

guessImpropers(x)

guessTopology(x, ...)

\method{guessTopology}{Atoms}(x, bonds = TRUE, angles = TRUE,
  dihedrals = TRUE, impropers = FALSE, ...)
}
\arguments{
\item{x}{an R object.}

\item{radius}{a numeric vector. Atomic radii used to defined bonded atoms.}

\item{safety}{a numeric value used to extend the atomic radii.}

\item{usePBC}{a logical value. Whether to use periodic boundary conditions to compute inter-atomic distances.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
\code{guessBonds} returns a data.frame with the following columns: 
  \describe{ \item{atm1,atm2}{integer vectors. Indices of atoms forming 
  bonds.} \item{h,k,l}{integer vectors. Miller indices of the image cell in 
  which \code{atm2} is located assuming \code{atm1} to be locate in the 
  original cell.} }
  
  \code{guessAngles} returns a data.frame with the following columns: 
  \describe{ \item{atm1,atm2,atm3}{integer vectors. Indices of atoms 
  forming bond angles.} }
  
  \code{guessDihedrals} returns a data.frame with the following columns: 
  \describe{ \item{atm1,atm2,atm3,atm4}{integer vectors. Indices of atoms
  forming dihedral angles.} }
  
  \code{guessTopology} returns an object of class \sQuote{Atoms} containing 
  the guessed topology.
}
\description{
Guess the topology of a molecular system.
}
\details{
\code{guessBonds} compairs inter-atomic distances of the current frame to 
bounding distances (computed from atomic radii) to determine which atoms are 
bonded.\cr\code{guessAngles} searches for connected bonds to determine bond 
angles.\cr\code{guessDihedrals} searches for overlapping bond angles to 
determine dihedral angles.\cr\code{guessTopology} calls the previews 
functions to guess the full topology of the system.
}

