% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure}
\alias{measure}
\alias{measure.Atoms}
\title{Measure Bonds Angles or Dihedrals}
\usage{
measure(x, atm1, atm2, atm3, atm4)

\method{measure}{Atoms}(x, atm1, atm2 = NULL, atm3 = NULL, atm4 = NULL)
}
\description{
Measure Bonds Angles or Dihedrals
}

