% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natom.R
\name{natom}
\alias{natom}
\title{Number Of Atoms}
\usage{
natom(x)
}
\arguments{
\item{x}{an R object}
}
\value{
an integer. The number of atoms.
}
\description{
Get the number of atoms stored inside an objet.
}
\details{
The S4 method for class \sQuote{Atoms} count the number of rows of slot
\code{atoms}.
}
\seealso{
\code{\link{Atoms-class}}
}

