% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nframe.R
\name{nframe}
\alias{nframe}
\title{Number Of Frames}
\usage{
nframe(x)
}
\arguments{
\item{x}{an R object}
}
\value{
an integer. The number of frames.
}
\description{
Get the number of frames stored inside an objet.
}
\details{
The S4 method for class \sQuote{Atoms} count the number of columns of slot
\code{a}.
}
\seealso{
\code{\link{Atoms-class}}
}

