% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc.R
\docType{methods}
\name{pbc}
\alias{pbc}
\alias{pbc,Atoms-method}
\alias{pbc<-}
\alias{pbc<-,Atoms-method}
\title{pbc}
\usage{
pbc(x)

\S4method{pbc}{Atoms}(x)

pbc(x) <- value

\S4method{pbc}{Atoms}(x) <- value
}
\description{
pbc
}

