% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbcBox.R
\name{pbcBox}
\alias{pbcBox}
\alias{pbcBox.Atoms}
\alias{pbcBox.matrix}
\title{draw PBC box}
\usage{
pbcBox(x, ...)

\method{pbcBox}{matrix}(x, ...)

\method{pbcBox}{Atoms}(x, ...)
}
\description{
draw PBC box
}

