% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/planNormals.R
\name{.planNormalsAtoms}
\alias{.planNormalsAtoms}
\alias{planNormals}
\alias{planNormals.Atoms}
\title{planNormals}
\usage{
.planNormalsAtoms(x, atm1, atm2, atm3)

planNormals(x, ...)

\method{planNormals}{Atoms}(x, atm1, atm2 = NULL, atm3 = NULL, ...)
}
\description{
planNormals
}

