% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rdf.R
\name{.rdfAtoms}
\alias{.rdfAtoms}
\alias{rdf}
\alias{rdf.Atoms}
\title{rdf}
\usage{
.rdfAtoms(x, sel1, sel2, resnumb, cutoff, interval, type)

rdf(x, ...)

\method{rdf}{Atoms}(x, sel1 = NULL, sel2 = NULL, resnumb = x$resnumb,
  cutoff = 20, interval = 0.02, type = "xyz")
}
\description{
rdf
}

