% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsd.R
\name{rmsd}
\alias{rmsd}
\alias{rmsd.Atoms}
\title{Root Mean Square Deviation}
\usage{
rmsd(x, y, ...)

\method{rmsd}{Atoms}(x, y, usePBC = TRUE)
}
\arguments{
\item{x}{an R object for which to compute the rmsd.}

\item{y}{the reference used to compute the rmsd. See details.}

\item{usePBC}{a logical value. Whether to use PBC to compute inter-atomic
distances.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
Compute the root mean square deviation between a structure and a reference
structure.
}
\details{
\code{y} can be an object of class \sQuote{Atoms} or an integer indicating a
frame of \code{x} used as a reference.

\code{x} and \code{y} must have the same lattice vectors and PBC.
}

