% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superCell.R
\name{superCell}
\alias{superCell}
\alias{superCell.Atoms}
\title{superCell}
\usage{
superCell(x, ...)

\method{superCell}{Atoms}(x, aInds = 0L, bInds = 0L, cInds = 0L,
  current = NULL)
}
\description{
superCell
}

