% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoDictionary.R
\name{topoDictionary}
\alias{topoDict}
\alias{topoDict.Atoms}
\alias{topoDictionary}
\title{Topology Dictionary}
\usage{
topoDict(x, atmprop, unique)

\method{topoDict}{Atoms}(x, atmprop = "atmtype", unique = TRUE)
}
\description{
Topology Dictionary
}

