% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topology.R
\name{topology}
\alias{topology}
\alias{topology.Atoms}
\alias{topology<-}
\alias{topology<-.Atoms}
\title{topology}
\usage{
topology(x)

\method{topology}{Atoms}(x)

topology(x) <- value

\method{topology}{Atoms}(x) <- value
}
\description{
topology
}

