% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topologyGettersAndSetters2.R
\name{topologyGettersAndSetters}
\alias{angles}
\alias{angles.Atoms}
\alias{angles<-}
\alias{angles<-.Atoms}
\alias{bonds}
\alias{bonds.Atoms}
\alias{bonds<-}
\alias{bonds<-.Atoms}
\alias{dihedrals}
\alias{dihedrals.Atoms}
\alias{dihedrals<-}
\alias{dihedrals<-.Atoms}
\alias{impropers}
\alias{impropers.Atoms}
\alias{impropers<-}
\alias{impropers<-.Atoms}
\alias{topologyGettersAndSetters}
\title{Extract or Replace Topology Components}
\usage{
bonds(x, ...)

angles(x, ...)

dihedrals(x, ...)

impropers(x, ...)

bonds(x) <- value

angles(x) <- value

dihedrals(x) <- value

impropers(x) <- value

\method{bonds}{Atoms}(x, atm1 = NA, atm2 = NA, atmprop = x$atmtype,
  nwise = TRUE, ...)

\method{angles}{Atoms}(x, atm1 = NA, atm2 = NA, atm3 = NA,
  atmprop = x$atmtype, nwise = TRUE, ...)

\method{dihedrals}{Atoms}(x, atm1 = NA, atm2 = NA, atm3 = NA, atm4 = NA,
  atmprop = x$atmtype, nwise = TRUE, ...)

\method{impropers}{Atoms}(x, atm1 = NA, atm2 = NA, atm3 = NA, atm4 = NA,
  atmprop = x$atmtype, nwise = TRUE, ...)

\method{bonds}{Atoms}(x) <- value

\method{angles}{Atoms}(x) <- value

\method{dihedrals}{Atoms}(x) <- value

\method{impropers}{Atoms}(x) <- value
}
\arguments{
\item{x}{an R object}

\item{value}{a data.frame containing by column \code{atm1}, \code{atm2},
\code{atm3} and \code{atm4} indices used to set bonds, bond angles,
dihedral angles or improper angles. Columns must be named \code{atm1},
\code{atm2}, \code{atm3} and \code{atm4}.}

\item{atm1, atm2, atm3, atm4}{atomic properties of atoms forming bonds, bond
angles, dihedral angles or improper angles.}

\item{atmprop}{a vector of atomic properties used for matching \code{atm1},
\code{atm2}, \code{atm3} and \code{atm4}.}

\item{nwise}{a logical value. Whether to consider \code{atm1} and
\code{atm2}/\code{atm3}/\code{atm4} as pair/triplet/quadruplet or not.}
}
\value{
\code{bonds}, \code{angles}, \code{dihedrals} and \code{impropers}
  return a data.frame containing the indices of the atoms forming specific
  bonds, bond angles, dihedral angles, improper angles.

  \code{bonds<-}, \code{angles<-}, \code{dihedrals<-} and \code{impropers<-}
  return an object of the same class as \code{x} with updated topology.
}
\description{
Extract or Replace bonds, bond angles, dihedral angles and improper angles.
}

