% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{trajectory}
\alias{trajectory}
\title{trajectory}
\usage{
trajectory(x)
}
\arguments{
\item{x}{an R object from/to which to get/set the trajectory}

\item{value}{an R object containing the trajectory used for setting \code{x}}
}
\description{
Get or set the lattice parameters, atomic poistions, velocities and forces of
an object.
}

