% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectOrientation.R
\name{vectOrientation}
\alias{vectOrientation}
\alias{vectOrientation.Atoms}
\title{Vectors Orientation}
\usage{
vectOrientation(x, ...)

\method{vectOrientation}{Atoms}(x, V1, V2, pairwise = TRUE, ...)
}
\arguments{
\item{V1}{a data.frame containing pairs/triplets of atom indexes defining a 
set of distance/normal vectors.}

\item{V2}{Same as V1 or a numeric vector of length 3 indicating a single 
reference vector for the calculation of the angles. See details.}

\item{pairwise}{a logical value. When both V1 and V2 are data.frame, 
indicates wheither to compute angles pairwise or not.}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
A 2D or 3D numeric array. The last dimension span over the frames of 
  the trajectory.
}
\description{
Calculate the angles between two sets of vectors (Distance vectors or plan 
normals).
}
\details{
When \code{V1} and \code{V2} are data.frame, they must contain columns named 
\code{atm1}, \code{atm2} and optionally \code{atm3}. They must all be integer
vectors containing atom indexes defining the vectors used for the analysis. 
When only \code{atm1} and \code{atm2} are specified, distances vectors are 
used. When \code{atm3} is also specified, plan normals are used instead. In
this latter case, plan normals are calculated as the cross product between
two normalized distance vectors d12 (defined by \code{atm1} and \code{atm2})
and d13 (defined by \code{atm1} and \code{atm3}).

When \code{V2} is missing, the angles between V1 and itself are calculated 
(lower-triangle of the (V1, V1) angles matrix). When \code{V2} is a numeric 
vector, the angles between V1 and this single reference vector are 
calculated. When both \code{V1} and \code{V2} are data.frame, the angles 
between V1 and V2 are calculated (a (N1,N2) matrix for each frame is 
calculated).
}

