% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{view}
\alias{view}
\alias{view.Atoms}
\title{Visualize atomic structure}
\usage{
view(x, ...)

\method{view}{Atoms}(x, drawing = "lines", color = NULL, radius = NULL,
  add = FALSE, FOV = 0, windowRect = c(0, 0, 800, 600),
  userMatrix = diag(4), pbc.box = NULL, ...)
}
\arguments{
\item{\dots}{rgl.materials}
}
\description{
Visualize atomic structure
}

