% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/wrap.R
\name{.wrapAtoms}
\alias{.wrapAtoms}
\alias{wrap}
\alias{wrap.Atoms}
\title{wrap}
\usage{
.wrapAtoms(x, resnumb, multi)

wrap(x, ...)

\method{wrap}{Atoms}(x, resnumb = x$resnumb, multi = TRUE, ...)
}
\description{
wrap
}

