% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbc.R
\docType{methods}
\name{pbc}
\alias{pbc}
\alias{pbc,Atoms-method}
\alias{pbc<-}
\alias{pbc<-,Atoms-method}
\title{Periodic Boundary Conditions}
\usage{
pbc(x)

\S4method{pbc}{Atoms}(x)

pbc(x) <- value

\S4method{pbc}{Atoms}(x) <- value
}
\arguments{
\item{x}{an R object fo which to get or set PBC.}

\item{value}{a logical vector of length 3. PBC along a-, b- and c-axis.}
}
\value{
\code{pbc} returns a logical vector of length 3 indicating wheither
  PBC are define respectively along the a-, b- and c-axis of the system.
}
\description{
Get or set the periodic boundary conditions (PBC) of an object.
}

